% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2earth.R
\name{plot2earth}
\alias{plot2earth}
\title{plot2earth}
\usage{
plot2earth(df, longitude, latitude, date.time, labels, size, colour,
  colour.scale = plotKML::SAGA_pal[[1]],
  shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png",
  open = TRUE, kmz = FALSE, file.name = "tweets2earth.kml",
  folder.name = getwd(), ...)
}
\arguments{
\item{df}{Required. data.frame of tweets to plot.}

\item{longitude}{Required. Column holding longitude.}

\item{latitude}{Required. Column holding latitude.}

\item{date.time}{Required. Column including creation time of tweets; 
\code{POSIXct} format}

\item{labels}{Required. Column holding labels.}

\item{size}{Required. Column holding desired size of points.}

\item{colour, }{optional. Column name that determines colours of points.}

\item{colour.scale, }{optional, defaults to \code{plotKML::SAGA_pal[[1]]}. 
can be pallettes from R_pal (\code{data(R_pal)}), SAGA_pal (see default), 
from \code{RColorBrewer} (see example) or a vector of colors (hex) 
i.e.(\code{c("#EE6A50", "#E8F9FE", "#6495ED")})}

\item{shape, }{optional. Shape to use for points. Defaults to 
\code{http.../icon18.png}}

\item{open, }{optional. Whether to open file in Google Earth. 
Defaults to \code{TRUE}.}

\item{kmz}{Whether to also save the file as \code{.kmz}, defaults to 
\code{FALSE}}

\item{file.name}{Name of kml or kmz file, defaults to \code{tweets2earth.kml}}

\item{folder.name}{Name of folder where to save the kml or kmz file, 
defaults to \code{getwd}}

\item{...}{Arguments to pass to \code{kml} function. 
See \code{\link[plotKML]{kml}}}
}
\value{
Writes kml file and, optionally, opens it in Google Earth 
(\code{open = TRUE}).
}
\description{
Plots tweets location and created time on Google Earth.
}
\details{
Takes data.frame of tweets and plots coordinates on Google Earth, also 
includes time variable (\code{date.time}). Choose \code{size}, 
\code{colour} and \code{labels} of your liking.
}
\examples{

\dontrun{
library(twitteR)

# setup oauth 
setup_twitter_oauth(consumer_key, consumer_secret, access_token=your_token, 
                    access_secret=your_secret)

# search twitter
tweets <- searchTwitter("rstats", n = 2000, lang= "en",
                        geocode="34.052235,-118.243683,100mi")

# unlist
tw.df <- do.call("rbind", lapply(tweets, as.data.frame))

plot2earth(tw.df, "longitude", "latitude", "created", 
           "screenName", "retweetCount", "favorited", 
           colour.scale = colour.scale = RColorBrewer::brewer.pal(8, "BuPu"),
           file.name = "rstats.kml")
}

}
\author{
John Coene, \email{jcoenep@gmail.com}
}
\seealso{
\\code{\link{stream2earth}}
}

