% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tweet2Earth.R
\name{plot2Earth}
\alias{plot2Earth}
\title{\code{plot2Earth}}
\usage{
plot2Earth(df, longitude, latitude, date_time, labels, size, colour,
  colour_scale = SAGA_pal[["SG_COLORS_GREEN_BLACK"]],
  shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png",
  open = TRUE, ...)
}
\arguments{
\item{df,}{require. data.frame of tweets to plot.}

\item{longitude}{Required. Column holding longitude.}

\item{latitude}{Required. Column holding latitude.}

\item{date_time}{Required. Column including creation time of tweets; POSIXct format}

\item{labels}{Required. Column holding labels.}

\item{size}{Required. Column holding desired size of points.}

\item{colour,}{optional. Column name that determines colours of points.}

\item{colour_scale,}{optional, defaults to SAGA_pal[["SG_COLORS_GREEN_BLACK"]]. Pallettes from R_pal and SAGA_pal. data(R_pal)}

\item{shape,}{optional. Shape to use for points.}

\item{open,}{optional. Whether to open file in Google Earth. Defaults to TRUE.}

\item{...}{Arguments to pass to kml() function.}
}
\value{
Writes kml files named "tweets2Earth" in working directory and opens file in Google Earth depending on open param.
}
\description{
Plots tweets location and created time on Google Earth.
}
\details{
Takes data.frame of tweets and plots coordinates on Google Earth, also includes created time. Choose size, color and labels of your liking.
}
\examples{
\dontrun{
library(twitteR)

# setup oauth
setup_twitter_oauth(consumer_key, consumer_secret, access_token=your_token, access_secret=your_secret)

# search twitter
tweets <- searchTwitter("searchTerm", n = 2000, lang= "en")

# unlist
data_frame <- do.call("rbind", lapply(tweets, as.data.frame))

# inspect
names(data_frame)
[1] "text"          "favorited"     "favoriteCount" "replyToSN"     "created"
[6] "truncated"     "replyToSID"    "id"            "replyToUID"    "statusSource"
[11] "screenName"    "retweetCount"  "isRetweet"     "retweeted"     "longitude"
[16] "latitude"

plot2Earth(data_frame, "longitude", "latitude", "created", "screenName", "retweetCount", "favorited")
}
}

