% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tweet2Earth.R
\name{stream2Earth}
\alias{stream2Earth}
\title{\code{stream2Earth}}
\usage{
stream2Earth(search_term, oauth, labels, size, duration = 1, window = 5,
  shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png", ...)
}
\arguments{
\item{search_term,}{required. Search to use in API call.}

\item{oauth,}{required. Your oauth token as used in streamR package. See ROauth package.}

\item{labels,}{requried. Labels to use on kml.}

\item{size,}{required. Size of shape.}

\item{duration,}{requried. Number of days to stream, defaults to 1.}

\item{window,}{required. Duration of the streaming window in seconds. Defaults to 5.}

\item{shape,}{required. Shape to use on Google Earth, has default.}

\item{...}{Optional arguments to pass to streamR filterStream function.}
}
\value{
Opens Google Earth and plots tweet location and created time in near-real time.
}
\description{
Streams tweets' location and time to Google Earth.
}
\details{
Streams location and created time onto Google Earth.
}
\note{
labels, size parameters must be inherited from tweets filterStream function.

Available labels and size parameters are;
names(filterStream.tweets)
[1] "text"                      "retweet_count"
[3] "favorited"                 "truncated"
[5] "id_str"                    "in_reply_to_screen_name"
[7] "source"                    "retweeted"
[9] "created_at"                "in_reply_to_status_id_str"
[11] "in_reply_to_user_id_str"   "lang"
[13] "listed_count"              "verified"
[15] "location"                  "user_id_str"
[17] "description"               "geo_enabled"
[19] "user_created_at"           "statuses_count"
[21] "followers_count"           "favourites_count"
[23] "protected"                 "user_url"
[25] "name"                      "time_zone"
[27] "user_lang"                 "utc_offset"
[29] "friends_count"             "screen_name"
[31] "time"
}
\examples{
\dontrun{
# OAuth setup - from App (http://apps.twitter.com/)
requestURL <- "https://api.twitter.com/oauth/request_token"
accessURL <- "https://api.twitter.com/oauth/access_token"
authURL <- "https://api.twitter.com/oauth/authorize"
my_oauth <- OAuthFactory$new(consumerKey = "your_consumer_key", consumerSecret = "your_consumer_secret",
                             requestURL = requestURL, accessURL = accessURL, authURL = authURL)

# download cacert.pem
download.file(url="http://curl.haxx.se/ca/cacert.pem", destfile="cacert.pem")

# Register OAuth
my_oauth$handshake(cainfo = system.file("CurlSSL", "cacert.pem", package = "RCurl"))

#Run stream for 2 days with windows of 7 seconds looking for keyword #rstats
stream2Earth("#rstats", oauth, "screen_name", "retweet_count", duration = 2, window = 7)
}
}

